refactor: 💡 변경 Enum 에러 출력 텍스트 제외하고 모두 삭제 및 이름 수정
에러 텍스트는 여기저기서 써야하니 남겨두고, 나머지 텍스트는 유지할 필요가 딱히 없을 것 같다.
