+++ src/main/java/constant/Text.java
@@ -1,0 +1,17 @@
+package constant;
+
+public enum Text {
+    ERROR("[ERROR] ")
+    ;
+
+    final String content;
+
+    Text(String content) {
+        this.content = content;
+    }
+
+    @Override
+    public String toString() {
+        return this.content;
+    }
+}
