+++ src/main/java/lotto/WinLotto.java
@@ -1,6 +1,6 @@
 package lotto;
 
-import constant.LottoText;
+import constant.Text;
 
 import java.util.List;
 
@@ -16,7 +16,7 @@ public class WinLotto extends Lotto {
     private void validate(List<Integer> numbers, int bonusNumber) {
         if (numbers.contains(bonusNumber)) {
             throw new IllegalArgumentException(
-                    LottoText.ERROR + "보너스 번호가 당첨 번호와 중복되었습니다.");
+                    Text.ERROR + "보너스 번호가 당첨 번호와 중복되었습니다.");
         }
     }
